put /usr/lib/postgresql into a separate volume
add command line --createdb that takes a url to download a .osm.pbf file
from, then imports it to pgsql.
add command line --threads that specifies the number of threads to pass
to the database import process.
add ability to autodetect number of cores available, and default to
using number of cores * 2 for threads